# docs/swagger.yaml
openapi: 3.0.0
info:
  title: FixNearby API Documentation
  description: API endpoints specification for local service discovery, booking, and workers.
  version: 1.0.0
paths:
...
  /api/bookings:
    post:
      summary: Create Booking
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                workerId:
                  type: string
                service:
                  type: string
                scheduledTime:
                  type: string
                  format: date-time
                durationHours:
                  type: number
                address:
                  type: string
                price:
                  type: number
      responses:
        '201':
          description: Booking created
    get:
      summary: Get User/Worker Bookings
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of bookings
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT